# docs/swagger.yaml
openapi: 3.0.3
info:
  title: LeetCode Clone API
  description: |-
    Полнофункциональный бэкэнд API для платформы программирования, подобной LeetCode. Предоставляет функции для управления задачами, пользователями, комментариями, рейтингами и файлами.
    
    ## Authentication
    Для большинства конечных точек требуется аутентификация JWT. Включите токен в заголовок Authorization.:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## User Roles
    - **user**: Базовый доступ, можно комментировать и оценивать задачи
    - **interviewer**: Может создавать и редактировать задачи  
    - **admin**: Полный доступ ко всем функциям
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, interviewer, admin]
          example: "user"
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
      required:
        - id
        - username
        - email
        - role
    Problem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Two Sum"
        description:
          type: string
          example: "Given an array of integers nums and an integer target..."
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        examples:
          type: array
          items:
            type: object
            properties:
              input:
                type: string
                example: "nums = [2,7,11,15], target = 9"
              output:
                type: string
                example: "[0,1]"
              explanation:
                type: string
                example: "Because nums[0] + nums[1] == 9"
        tags:
          type: array
          items:
            type: string
          example: ["array", "hash-table"]
        averageRating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.5
        ratingCount:
          type: integer
          example: 10
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
      required:
        - id
        - title
        - description
        - difficulty
    ProblemCreate:
      type: object
      properties:
        title:
          type: string
          example: "Two Sum"
        description:
          type: string
          example: "Given an array of integers nums and an integer target..."
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        examples:
          type: array
          items:
            type: object
            properties:
              input:
                type: string
              output:
                type: string
              explanation:
                type: string
        tags:
          type: array
          items:
            type: string
          example: ["array", "hash-table"]
      required:
        - title
        - description
        - difficulty
    Comment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        content:
          type: string
          example: "Great problem for beginners!"
        userId:
          type: integer
          example: 1
        problemId:
          type: integer
          example: 1
        parentId:
          type: integer
          nullable: true
          example: null
        user:
          $ref: '#/components/schemas/User'
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
    CommentCreate:
      type: object
      properties:
        content:
          type: string
          example: "This is a helpful comment"
        problemId:
          type: integer
          example: 1
        parentId:
          type: integer
          nullable: true
          example: null
      required:
        - content
        - problemId
...
    File:
      type: object
      properties:
        id:
          type: integer
          example: 1
        filename:
          type: string
          example: "solution-123456789.pdf"
        originalName:
          type: string
          example: "solution.pdf"
        path:
          type: string
          example: "/uploads/solution-123456789.pdf"
        problemId:
          type: integer
          example: 1
        uploadedBy:
          type: integer
          example: 1
        fileSize:
          type: integer
          example: 1024
        mimeType:
          type: string
          example: "application/pdf"
        description:
          type: string
          example: "Solution explanation"
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T10:00:00Z"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          example: "password123"
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          example: "password123"
        role:
          type: string
          enum: [user, interviewer, admin]
          default: "user"
          example: "user"
      required:
        - username
        - email
        - password
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message description"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          additionalProperties: true
    PaginatedResponse:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
    CommentsResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
...
    FilesResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            files:
              type: array
              items:
                $ref: '#/components/schemas/File'
...
  parameters:
    ProblemId:
      name: problemId
      in: path
      required: true
      schema:
        type: integer
      description: ID of the problem
    CommentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID of the comment
...
    FileId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID of the file
...
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error in request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      summary: Get current user information
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Logout user
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"
...
  /problems/{problemId}:
    get:
      summary: Get problem by ID
      tags: [Problems]
      parameters:
        - $ref: '#/components/parameters/ProblemId'
      responses:
        '200':
          description: Problem details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update problem
      tags: [Problems]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProblemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemCreate'
      responses:
        '200':
          description: Problem updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete problem
      tags: [Problems]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProblemId'
      responses:
        '204':
          description: Problem deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /comments:
    post:
      summary: Create a new comment
      tags: [Comments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/ValidationError'
  /comments/problems/{problemId}:
    get:
      summary: Get comments for a problem
      tags: [Comments]
      parameters:
        - $ref: '#/components/parameters/ProblemId'
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsResponse'
  /comments/{id}:
    put:
      summary: Update comment
      tags: [Comments]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Can only edit your own comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete comment
      tags: [Comments]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /files/upload:
    post:
      summary: Upload a file for a problem
      tags: [Files]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                problemId:
                  type: integer
                  description: ID of the problem
                description:
                  type: string
                  description: File description
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "File uploaded successfully"
                  file:
                    $ref: '#/components/schemas/File'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          description: No file uploaded or problemId missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /files/problem/{problemId}:
    get:
      summary: Get files for a problem
      tags: [Files]
      parameters:
        - $ref: '#/components/parameters/ProblemId'
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilesResponse'
  /files/{id}/download:
    get:
      summary: Download a file
      tags: [Files]
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{id}:
    delete:
      summary: Delete a file
      tags: [Files]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204':
          description: File deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...